Split task and business logic of swapFile/swapSlice operations
