+++ internal/pkg/service/buffer/condition/import.go
@@ -3,9 +3,19 @@ package condition
 import (
 	"context"
 	"fmt"
+	"strings"
 	"time"
 
+	"github.com/keboola/keboola-as-code/internal/pkg/log"
+	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/file"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/store/key"
+	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/usererror"
+	"github.com/keboola/keboola-as-code/internal/pkg/service/common/rollback"
+	"github.com/keboola/keboola-as-code/internal/pkg/service/common/task"
+)
+
+const (
+	fileSwapTaskType = "file.swap"
 )
 
 func (c *Checker) shouldImport(ctx context.Context, now time.Time, sliceKey key.SliceKey, uploadCredExp time.Time) (ok bool, reason string, err error) {
@@ -37,4 +47,35 @@ func (c *Checker) shouldImport(ctx context.Context, now time.Time, sliceKey key.
 	// Evaluate import conditions
 	ok, reason = evaluate(export.ImportConditions, now, sliceKey.FileKey.OpenedAt(), fileStats.Total)
 	return ok, reason, nil
+}
+
+func (c *Checker) startSwapFileTask(fileManager *file.AuthorizedManager, fileKey key.FileKey, reason string) error {
+	return c.tasks.StartTaskOrErr(task.Config{
+		Type: fileSwapTaskType,
+		Key: task.Key{
+			ProjectID: fileKey.ProjectID,
+			TaskID: task.ID(strings.Join([]string{
+				fileKey.ReceiverID.String(),
+				fileKey.ExportID.String(),
+				fileKey.FileID.String(),
+				fileSwapTaskType,
+			}, "/")),
+		},
+		Context: func() (context.Context, context.CancelFunc) {
+			return context.WithTimeout(context.Background(), time.Minute)
+		},
+		Operation: func(ctx context.Context, logger log.Logger) (result task.Result) {
+			defer usererror.CheckAndWrap(&result.Error)
+
+			rb := rollback.New(logger)
+			defer rb.InvokeIfErr(ctx, &result.Error)
+
+			logger.Infof(`closing file "%s": %s`, fileKey, reason)
+			if err := fileManager.SwapFile(ctx, rb, fileKey); err != nil {
+				return task.ErrResult(err)
+			}
+
+			return task.OkResult("new file created, the old is closing")
+		},
+	})
 }
